fix(parsing): Make Parser.parse() a classmethod
Originally, this code failed to run, because Parser.parse() expected an instance of the Parser
object to exist before being able to parse a file. This change enables Parser.parse() to work
correctly without requiring the user to first create a Parser object. Parser objects are temporary,
and only exist to parse tokens created by the Lexer. As such, it makes sense for Parser.parse() to
create the temporary object, run it, and return the results to the caller.